gpu: mali: Bypass check for HW errata 1717279
r25p0 makes the implementation of Mali-G78 (aka Borr)
erratum 1717279 compulsory. This only applies to the
version of Borr that is in A0. B0's GPU version has
this erratum fixed.

We should still consider implementing the workaround
for this erratum, but for now this CL bypasses the
check that fails the GPU device probe if the WA is
not configured.

Bug: 162413384
Bug: 159019992
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: Ic426b543d793a3b78cd264cfc52285dc285e4a6c