mali_kbase: mmu: Only MMU flush when the GPU is on
Currently the core driver logic avoids flushing and
invalidating an address space when the system is in
suspend. Since it is valid to have the GPU powered off
when the overall system is not suspended, we can
avoid the flush when just the GPU is powered down.

Without this, we will needlessly power up the GPU
to perform the flush, raising power usage.

Bug: 156057140
Bug: 163884347
Bug: 166462595
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I54c4169d1551880c14d0b79f38d4edae18fe1e31