mali_kbase: mem: Avoid unnecessary cache flushes
The Mali core driver has implemented a workaround to cope with
past buggy implementations of ION's begin_cpu_access and
end_cpu_access functions in past versions of the kernel.

In 4.19 we reduced the impact of this by only performing cache
operations when we detected that the buffer was cached.

In the ION implementations of these dma-buf operations in 5.4
it doesn't appear that any of the issues are still present.
The functions also perform an explicit check on the private
handle to determine whether the buffers are not cached in which
case they do not perform any cache operations.

In summary, it would appear that we need neither the workaround
nor the optimization to skip cache ops on uncached buffers as
this is all done in the new ION implementation.

Bug: 151952998
Bug: 167718192
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: Ib177097c32d3e97d4aa266b801497184b499b31e