mali_kbase: platform: dvfs: Support different ASV table sizes
Different spins of GS101 have different DVFS operating points
and to support this we have two distint DVFS tables in the GPU
devicetree entry.

Earlier versions of the Pixel Integration code required that
both DVFS tables were of the same size which was overly
restrictive. This change allows for each of the DVFS tables to
have their own size.

Bug: 177418501
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: Icddb05b40bc83318eab0747d6d8e2cb64f29d7e9