mali_kbase: platform: dvfs: Treat G3D clock as main freq
Treat the G3D (i.e. shader stack clock) as the primary clock for
the GPU. This is mainly due to the fact that the G3D clock is the
primary indicator of performance/power for a given operating
point in the DVFS table. See the bug for more rationale.

This also means that the G3D is the key value in the DVFS table.

Bug: 177418377
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I0ff2b0972afff140e40a80ef6fc24394ad4f7ff4