mali_kbase: platform: secure: Assert on correct lock
The code that we run when disabling secure mode on the GPU is
asserting that the wrong lock is held. While entering secure
mode should only happen when the hwaccess lock is held,
disabling secure mode should assert that the pm lock is held
instead due to the differences in the way the Mali driver
enters and exits secure mode.

Bug: 178425200
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: Ic2f7f90760135357a3c09bb56108d47c4de3da56