mali_kbase: platform: dvfs: Ensure clock updates on poweron.
The DVFS application for gpu only happens when GPU is on.
The DVFS level is currently assessed when utilization event comes in.
If utilization or policy changed the level when GPU was off, it could
have been quite some time until it was actually applied. This
particularly affected low duty cycle applications, such as UI
interactions. Higher duty-cycle applications would have more window to
apply DVFS changes as GPU was on more.

Fixes: 179395357
Test: boot, verify systrace timings
Change-Id: Ib3a186a69400012a53ad2a3074f977de8a56ae4b