mali_kbase: platform: dvfs: Increase freq of util calculation
Increases the frequency at which the core Mali driver evaluates the
current utilization of the GPU. This results in more frequency
evaluations of the operating point at which the GPU should run at
making it more responsive to changes to GPU load at the expense of
slightly higher driver overhead.

Note that this change also affects the absolute duration of hysteresis
values that are defined in terms of 'DVFS ticks' such as those
specified in the DVFS table in devicetree.

Bug: 179671040
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I5bbce1bcc969de412e55bc67bbfd7a088e9e3e09