mali_kbase: platform: dvfs: Fix bug with level locks
If an update to level lock resulted in a change in both
max and min levels for that lock, it was possible that
identifying that no effective lock was set would fail.

This also fixes a bug in the clock_info sysfs node which
would incorrectly assume a lock was set.

Bug: 179365671
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: Ia1dbb3fa9c04fee6bb95d9844ad99ad2c46b4b99