mali_kbase: mem: Don't snoop external memory
Currently when importing externally allocated (and owned)
memory we mark the mappings as being snoopable. This results
in snoop requests being issued for any access to these
regions.

Since the owners of these regions perform their own cache
operations before the GPU accesses them (or use uncached
memory), any snoop requests that the GPU raises are
superfluous, and only serve to increase power usage and
congestion on the memory fabric.

Bug: 151952998
Bug: 168113578
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I596a295444302754debc4a2634f94deb4f8fdccf