mali_kbase: platform: dvfs: Ensure level lock consistency
If a new max or min value is specified for a level lock that
conflicts with the current values in that level lock, reset the
old conflicting value(s).

For example, if the current lock has {min, max} set to {null, 200}
and a call is made to set the new min lock to 300, the resulting
values in the lock will be {300, null} since 300 > 200.

Prior to this change, this would have resulted in the lock being
set to {300, 300} which would cause issues when the min lock was
released as it would leave the max hanging higher than intended.

Bug: 188034128
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I3a27d11d959d6bd650624558b6015fe1b0370118