mali_kbase: mem: Prevent vma splits
When a subset of a vma region is unmapped, the kernel
attempts to split the vma region such that subset that
wasn't explicitly unmapped remains active. Due to the
way that kbase handles vma regions, they should never
be partially unmapped. A split therefore should only
be encounterd as a result of a bug in map or unmap
in the UMD.

This change implements the split callback and returns
-EINVAL to inform the kernel that it should not attempt
a split, and instead unmap the entire range instead.

It also raises a warning to indicate that there is
likely to be a bug in the upper levels of the driver
that should be investigated.

Bug: 190183863
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: Ieb0825667c43aaf517aa9e786553d03cd2ffe341