mali_kbase: platform: power: Correctly track GPU state
Ensure that the GPU top level power domain does not power down
when we only go idle so that GPU state is not lost. This allows us
to save cycles when there is new GPU work by not requiring the
core Mali KMD to regenerate state.

In order to do this, we track both GPU power domains and take
a reference to the GPU top level power domain when the SOC is not
suspended.

Bug: 191136938
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I015b3f3f2747a0b6c2f51609d07f0e475bfcac84