mali_kbase: platform: power: Correctly track GPU state
Ensure that the GPU top level power domain does not power down every
time the GPU goes idle so that GPU state is not lost. This allows us
to save cycles when there is new GPU work by not requiring the
core Mali KMD to regenerate state on each idle->active transition.

In order to do this, we track both GPU power domains and only power
down the GPU top level power domain after the GPU has remained idle
for a configurable duration. This duration is specified by the
`gpu_pm_autosuspend_delay` property in the GPU's device tree entry.

Bug: 191136938
Signed-off-by: Sidath Senanayake <sidaths@google.com>
Change-Id: I83f3d257638d2b18802dab725af93bab42fac856