mali_kbase: platform: revert pm.domain optionality
The PM domain being unavailable is a situation that should only occur
briefly during bringup. The workaround is easy enough to apply again
from scratch if needed, and since it's not a configuration we would
test, it would probably updating before the next bringup anyway. So the
value of keeping it around in the code in some form isn't worth the
additional complexity. We no longer need it now, so out it goes.

Bug: 190569515
Signed-off-by: Jesse Hall <jessehall@google.com>
Change-Id: I85e366e215c1e8089f00d65a1b56dc107983eaef