mali_kbase: platform: power: simplify gpu power management
The previous pixel power management code was based on an incorrect
understanding of how kernel power-domains mapped to hardware power
domains. While potential changes to the kernel power domains are being
investigated, this change switches to a very simple design to unblock
other work.

In this system we forego runtime pm and just control the frontend power
domains via direct exynso-pd calls. We only power off the frontend
domains during system suspend.

Power to the shader cores is not currently handled by any kernel power
domain, so we control it via a "backdoor" protocol with acpm. In this
change we also only power off shader cores during suspend, but a change
to do this whenever the gpu is idle will be coming shortly.

This is not the expected end state, just a simple but fully functional
starting point to build from.

Bug: 192307394
Signed-off-by: Jesse Hall <jessehall@google.com>
Change-Id: I6e926a27be68f1c198bcaf5621d79d02f61d3d5e