mali_kbase: platform: Re-add gs101 power management
This change allows building a gs101 gpu driver from this branch again,
re-adding the power management code that was (temporarily) rewritten for
gs201. This is just the first step: none of the differences have been
removed yet; CONFIG_SOC_* is used to pick the right code path for the
current build target.

Bug: 197597613
Signed-off-by: Jesse Hall <jessehall@google.com>
Change-Id: I69ad36a8b6993684390d6e011a10879d60a4c898