mali_kbase: turn on MALI_HOST_CONTROLS_SC_RAILS
HOST_CONTROLS_SC_RAILS is a device specific feature that will turn off
the shader cores in between frames for power savings.

It does this by turning off the rail after reception of an IRQ, and
then having firmware instruct userspace driver to trap CSGs to the
kernel so that power rails can be turned on before the next job.

Bug: 207028973
Bug: 201222260
Test: boot to camera.
Test: manual jank observations looked ok
Test: 16 hour astress run.
Test: Power rail savings observed.

Change-Id: I3928f1698f89026b5d00ae9f1e6f87f24f04d5f9
Signed-off-by: Kevin DuBois <kevindubois@google.com>