GPUCORE-33327 Send CSG error notification on MCU hang
This patchset introduces userspace error notifications for scheduled
CSGs when the MCU is unresponsive. Previously, the scheduler would only
notify userspace of CSG termination if the CSGs were suspending at the
time, and would otherwise reset the GPU without notifying userspace,
causing hangs in active applications. We noticed that this change
results in duplicate error notifications being sent to userspace
in certain conditions, but this does not cause any functional problems
as they merely indicate that the CSGs need to be recreated.

Bug: 232508335
Test: Manually trigger reset while SF is active
Change-Id: I2ba878f48023a76065f6f295d6360fcf683c8219
Signed-off-by: Jonglin Lee <jonglin@google.com>