GPUCORE-33799 Race between power policy & scheduler
When switching between any of the power policies and always_on policy,
there is a potential for a race condition as the scheduler is invoked
and a consistent state is expected while waiting for L2 to be off.

We now take the scheduler lock in power policy setting to avoid other
scheduler operations interfering.

Provenance: https://code.ipdelivery.arm.com/c/GPU/mali-ddk/+/4014
Bug: 224669116
Test: Toggle power policy to and from always_on in loop for 1hr
Change-Id: Ifc39fb4a237a5ccf312904bf8eb7956007b6f4dd
Signed-off-by: Jack Diver <diverj@google.com>