uci: make cfg parse on init optional
On newer androids it's not possible to parse the files before
a user process touches sdcard, so it's useless to try and
also adds a lot of unneeded logging too, when someone
does not use the cleanslate config apps at all.

On p6/pro not needed to parse it on init.

(cherry picked from commit ca35d0b730310c6294630d52300a1b392f5d1bfd)