mali_kbase: Ensure GPU L2 is up prior to MMU cmd
Fix a race condition where the thread issuing an MMU command is not the
owner of the first PM reference and the GPU is off. In that case, the
thread will not be blocked until the GPU L2 is on, and may issue an MMU
command before that, resulting in a hang.

Bug: 229473975
Bug: 229335764
Signed-off-by: Jack Diver <diverj@google.com>
Change-Id: I750928582c091553df1377e3a6c141cab5443cf4