mali_kbase: Ensure GPU L2 is up prior to MMU cmd
Fix races where the L2 has not finished powering up before we issue an
MMU command. It's possible that the power up be deferred until after a
scheduled power off, and in that case l2_desired may transiently be false.
Replacing the current wait, with one that explicitly waits for L2 power
up ensures that we don't issue a command for L2 power up has completed.

Bug: 229473975
Signed-off-by: Jack Diver <diverj@google.com>
Change-Id: Iab0337e23cef3bf10122bf669051c7e1062c7060