GPUCORE-37217 Restrict permissions for TLSTREAM interface
Permissions have been restricted for the interface to acquire
a file descriptor for the Timeline Stream. Unless the user process
is privileged, now at least one of these conditions must be
satisftied:

 * The kbase_unprivileged_global_profiling module parameter
   has been set to 1.
 * The user process has the CAP_SYS_ADMIN capability.
 * The user process has the CAP_PERFMON capability.

The parameter callbacks only allow enabling the feature
if it's disabled, but not disabling it once it is enabled.

Global profiling is allowed for unprivileged processes
when Kbase development support is not stripped.

The permissions for the debugfs entry of TLSTREAM have been
restricted, too, in order to prevent non-privileged processes
to access this entry even in systems where /sys/kernel/debug
is accessible to non-privileged processes (which is usually
not the case).

Change-Id: Ie5f113a8beea5b5ce5c1991b4c863ab8cb64c455
(cherry picked from commit ef2064d3e9cf8a351411bdb28baa1cc22b3e293c)
Bug: 272345974
Provenance: https://code.ipdelivery.arm.com/c/GPU/mali-ddk/+/4942
Signed-off-by: Guus Sliepen <gsliepen@google.com>