Revert "mali_kbase: mem: Prevent vma splits"
In the original bug, the protected memory imports via Base
were ignoring the actual size of the import that came back from
the kernel memory import routines. These resulted in errors as when
these imports were freed, the incorrect size was passed resulting in
only sub-regions of the original mapped range being unmapped resulting
in cases where the GPU and CPU VAs ended up being inconsistent.

A WAR was added to prevent VMA splits temporarily until a fix was provided
for the protected memory size mismatch. As a result of this fix this WAR is
no longer necessary. The consequences of this WAR is now resulting in failures
for the case when an application tries to call mprotect(restrictive) on a memory already
allocated and mmapped on by Vulkan API calls.

Vulkan alloc() invokes the cmem_heap_alloc() function, which for the general
case allocates some extra memory to fulfil the worse case alignment requirements.
As a result invoking mprotect on the partial user provided range always result
in VMA splits().

For further reference look at this article.
https://lwn.net/Articles/182847/

Bug 269535398

This reverts commit 6d1d889156e68493842f5bb18fc9aed74cc57454.

Change-Id: Ic5749fab2613d6495fd3669356697ff40bfafcb7