vh: sched: Count estimated util for both fg and bg groups
Currently, estimated util is added to a rq first, then calls the hook
function, trace_android_rvh_enqueue_task_fair, in which we add the
util_est to our vendor util if it is bg group. Because we get our fg
vendor util by subtracting the bg group util_est from rq util_est,
there could be a race that a big bg util_est is added to the rq but
not to the bg util group yet, and cpu frquency update is requested
at that time. In this case, we will get a high cpu frequency even
the bg group is throttled. To avoid it, we need to count both fg
and bg util_est.

Bug: 288527317
Change-Id: I83fce6e5acb4f56d441684201d5fb00f05f0faa1
Signed-off-by: Rick Yiu <rickyiu@google.com>
(cherry picked from commit e51f5b1e05afd783565d3828dbca91a964f9019b)