mali_kbase: restore hysteresis time.
An accidental change back to a lower hysteresis time as
a part of the R43P0 merge commit resulted in GPU
powering on and off more frequently and regressing previously
established perf. Restore the same to a value that was previously
determined and set in our downstream driver.

Bug 292157710
Bug 278174418

Change-Id: Id44337bc1fd672722a5edaf27aaca16f320b7ad0
Signed-off-by: Debarshi Dutta <debarshid@google.com>