mali_kbase: Use kthread for protm_event_worker
protm_event_worker is responsible for triggering protected mode entry,
but it currently runs in a default priority workqueue, which results
in occasional long stalls for protected mode work.

The workload of this task is very light, typically only running
for a few tens of microseconds each invocation, so converting
it to a kthread should have no significant impact on overall
system performance.

Bug: 297019294
Test: Manual Netflix PiP perfetto trace inspection
Change-Id: Ica9e744bd390ae6b88a701bf2d390f82937581d0