eh: Use interrupts to wait for compression completion
Rather than polling in the compression thread, it is faster and much more
efficient to wait for an interrupt to signal compression completion. This
was previously disabled for performance reasons because the compression
callback was processed directly in the compression IRQ's thread, which was
bottlenecked by the CPU that had the IRQ affined. With a separate thread
for compression however, this is not a problem.

Since it's easy to get a flood of redundant compression interrupts, extra
care is needed to make this as fast and efficient as possible. As such, the
compression interrupt is only enabled on demand when the compression thread
needs to wait, and synchronization between the compression thread and IRQ
handler is lock-less aside from the swaitqueue to ensure that both the
compression thread and IRQ handler are lean and fast.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit 1c3717b9a1801705a6cbb1ae34be220c5afe90e6)
(cherry picked from commit 0a53f55eec7fdf66bf1c92e8035c7811745d4aee)