kbuild: Build dtbs by default
It is possible for a dtbo to depend upon a dtb without actually having the
dependency correctly encoded. And then, if the bootloader happens to not
properly check for the existence of its matching dtb, catastrophe could
strike.

Without being set as a dependency for any dtbo, the B0 gs201 dtb isn't
built and thus doesn't get included in the final dtb image. If the
bootloader were robust, it would recognize that none of the dtbs installed
on the device match the device's hardware and thus prevent boot from
proceeding.

However, the bootloader isn't robust, and picks any of the available dtbs
when the correct dtb for the device is missing. If IPOP and A0 dtbs are
present, then cheetah MP will pick IPOP. If only A0 is present, then
cheetah MP will pick A0.

The correct dtb for cheetah/panther MP is B0, but the B0 dtb is only listed
as a dtb target, not as a dtbo dependency.

Build dtbs by default to ensure all enabled dtbs are built to fix this
without facing it again in the future.

Reported-by: freak07 <michalechner92@googlemail.com>
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit d62af1f7c395caadbe2f8b9c6a3dbbb2e970bec7)